+++ NEWS.md
@@ -96,3 +96,4 @@
 # eppasm 0.6.0
 
 - Explicitly model new diagnoses.
+- Add number of new diagnoses and ART initiations by age, sex, and CD4 category as model outputs.